refactor: seperate ItemModal component refactor: merge feature RemoteBoard -> Board component
