+++ juwonlee/board-project/src/components/Board/index.js
@@ -2,42 +2,9 @@ import { useState, useEffect } from 'react'
 import  './Board.css'
 import Header from './Header'
 import { BoardItem, NoItem } from './BoardItem'
-import Input from '../Form/Input'
+import { ItemModal, AddBtn } from '../Form/ItemModal'
 
 function Board() {
-  const items = [
-    {
-      id: crypto.randomUUID(),
-      seq: 1,
-      title: 'title 1',
-      content: 'content 1',
-      creator: 'creator 1',
-    },
-    {
-      id: crypto.randomUUID(),
-      seq: 2,
-      title: 'title 2',
-      content: 'content 2',
-      creator: 'creator 2',
-    },
-    {
-      id: crypto.randomUUID(),
-      seq: 3,
-      title: 'title 3',
-      content: 'content 3',
-      creator: 'creator 3',
-    },
-  ];
-
-  return (
-    <main>
-      <Header />
-      <ItemWrapper items={ items } />
-    </main>
-  );
-}
-
-function RemoteBoard() {
   const [items, setItems] = useState([]);
   useEffect(() => {
     loadItems().then((res) => setItems((res || []).map((item, index) => ({ seq: index+1, ...item }))))
@@ -131,46 +98,9 @@ function ItemWrapper({ items, onAddItem, onEditItem, onRemoveItem }) {
   );
 }
 
-function ItemModal({ title, selectedItem, actionText, itemHandler, isShow, onClose }) {
-  const item = { ...selectedItem() }
-
-  const onActionandler = (item) => {
-    itemHandler(item)
-    onClose(false)
-  }
-
-  return (
-    <div className={ `modal-wrapper ${isShow ? 'show' : ''}` }>
-      <div className='modal-content'>
-        <span className='close' onClick={() => onClose(false)}>&times;</span>
-        <p>{ title || '도구창' }</p>
-        <Input id='input_title' type='text' label='title' value={item?.title || ''} onChange={(e) => item.title = e.target.value}/>
-        <Input id='input_content' type='text' label='content' value={item?.content || ''} onChange={(e) => item.content = e.target.value}/>
-        <Input id='input_creator' type='text' label='creator' value={item?.creator || ''} onChange={(e) => item.creator = e.target.value}/>
-        <div className='panel'>
-          <button onClick={() => onActionandler(item)}>{ actionText || '확인' }</button>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function AddBtn({text, onClick}) {
-  const onClickHandler = () => {
-    onClick()
-    const itemList = document.querySelectorAll('.list-item') || []
-    const lastElement = itemList[itemList.length - 1]
-    if (lastElement) lastElement.scrollIntoView({ behavior: 'smooth', block: 'center' })
-  }
-
-  return (
-    <div className='add-btn' onClick={ onClickHandler }>{ text || 'Action' }</div>
-  )
-}
-
 const loadItems = async () => {
   const res = await fetch("http://heojh.iptime.org:8003/board");
   return await res.json();
 }
 
-export { Board, RemoteBoard };
+export default Board;
