+++ juwonlee/board-project/src/index.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import { RemoteBoard } from './components/Board';
+import Board from './components/Board';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <RemoteBoard />
+    <Board />
   </React.StrictMode>
 );
